// backup/backup-2025-05-21_13-08-02/src/types/supabase.ts
export type Database = {
// ...
  public: {
    Tables: {
      employees: {
        Row: {
          contact: string
          created_at: string | null
          hire_date: string
          id: string
          name: string
          position: string
          salary: number
          status: string
        }
        Insert: {
          contact: string
          created_at?: string | null
          hire_date: string
          id?: string
          name: string
          position: string
          salary: number
          status?: string
        }
        Update: {
          contact?: string
          created_at?: string | null
          hire_date?: string
          id?: string
          name?: string
          position?: string
          salary?: number
          status?: string
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}